+++ src/stcal/ramp_fitting/gls_fit.py
@@ -526,27 +526,22 @@ def gls_fit_single(ramp_data, gain_2d, readnoise_2d, max_num_cr, save_opt):
     group_time = ramp_data.group_time
     nframes_used = ramp_data.nframes
 
-    gdq_cube = ramp_data.groupdq
-    data_sect = ramp_data.data
-    input_var_sect = ramp_data.err
+    gdq = ramp_data.groupdq
+    data = ramp_data.data
     pixeldq = ramp_data.pixeldq
 
     jump_flag = ramp_data.flags_jump_det
     saturated_flag = ramp_data.flags_saturated
 
-    number_ints = data_sect.shape[0]
-    ngroups = data_sect.shape[1]
-    # number_rows = data_sect.shape[2]
-    # number_cols = data_sect.shape[3]
-
-    # imshape = (data_sect.shape[2], data_sect.shape[3])
+    number_ints = data.shape[0]
+    ngroups = data.shape[1]
 
     slope_int, slope_err_int, dq_int, temp_dq, slopes, sum_weight = \
-        create_integration_arrays(data_sect.shape)
+        create_integration_arrays(data.shape)
 
     # REFAC
     (intercept_int, intercept_err_int, pedestal_int, first_group, shape_ampl,
-        ampl_int, ampl_err_int) = create_opt_res(save_opt, data_sect.shape, max_num_cr)
+        ampl_int, ampl_err_int) = create_opt_res(save_opt, data.shape, max_num_cr)
 
     med_rates = None
     if ngroups == 1:
@@ -554,22 +549,25 @@ def gls_fit_single(ramp_data, gain_2d, readnoise_2d, max_num_cr, save_opt):
 
     # We'll propagate error estimates from previous steps to the
     # current step by using the variance.
-    input_var_sect = input_var_sect ** 2
+    input_var = ramp_data.err**2
 
     # Convert the data section from DN to electrons.
-    data_sect *= gain_2d
+    data *= gain_2d
 
-    # loop over data integrations
     for num_int in range(number_ints):
         ramp_data.current_integ = num_int
+        gdq_cube = gdq[num_int, :, :, :]
+        data_cube = data[num_int, :, :, :]
+        input_var_sect = input_var[num_int, :, :, :]
+
         if save_opt:
-            first_group[:, :] = data_sect[num_int, 0, :, :].copy()
+            first_group[:, :] = data[num_int, 0, :, :].copy()
 
         (intercept_sect, intercept_var_sect, slope_sect,
          slope_var_sect, cr_sect, cr_var_sect) = determine_slope(
              ramp_data,
-             data_sect[num_int, :, :, :], input_var_sect[num_int, :, :, :],
-             gdq_cube[num_int, :, :, :], readnoise_2d, gain_2d, frame_time, group_time,
+             data_cube, input_var_sect, gdq_cube,
+             readnoise_2d, gain_2d, frame_time, group_time,
              nframes_used, max_num_cr, saturated_flag, jump_flag, med_rates)
 
         slope_int[num_int, :, :] = slope_sect.copy()
@@ -579,7 +577,7 @@ def gls_fit_single(ramp_data, gain_2d, readnoise_2d, max_num_cr, save_opt):
             slope_var_sect[v_mask] = utils.LARGE_VARIANCE
 
             # Also set a flag in the pixel dq array.
-            temp_dq[:, :][v_mask] = ramp_data.flags_unreliable_slope
+            temp_dq[v_mask] = ramp_data.flags_unreliable_slope
             del v_mask
 
         # If a pixel was flagged (by an earlier step) as saturated in
@@ -587,8 +585,7 @@ def gls_fit_single(ramp_data, gain_2d, readnoise_2d, max_num_cr, save_opt):
         # Note:  save s_mask until after the call to utils.gls_pedestal.
         s_mask = (gdq_cube[0] == saturated_flag)
         if s_mask.any():
-            # TODO The dimensions of s_mask are larger than temp_dq
-            temp_dq[:, :][s_mask] = ramp_data.flags_unreliable_slope
+            temp_dq[s_mask] = ramp_data.flags_do_not_use
         slope_err_int[num_int, :, :] = np.sqrt(slope_var_sect)
 
         # We need to take a weighted average if (and only if) number_ints > 1.
@@ -606,6 +603,7 @@ def gls_fit_single(ramp_data, gain_2d, readnoise_2d, max_num_cr, save_opt):
             pedestal_int[num_int, :, :] = utils.gls_pedestal(
                 first_group[:, :], slope_int[num_int, :, :],
                 s_mask, frame_time, nframes_used)
+            del s_mask
 
             ampl_int[num_int, :, :, :] = cr_sect.copy()
             ampl_err_int[num_int, :, :, :] = np.sqrt(np.abs(cr_var_sect))
@@ -614,13 +612,12 @@ def gls_fit_single(ramp_data, gain_2d, readnoise_2d, max_num_cr, save_opt):
         # pixels
         pixeldq_sect = pixeldq[:, :].copy()
         dq_int[num_int, :, :] = utils.dq_compress_sect(
-            ramp_data, num_int, gdq_cube[num_int, :, :, :], pixeldq_sect).copy()
+            ramp_data, num_int, gdq_cube, pixeldq_sect).copy()
 
         dq_int[num_int, :, :] |= temp_dq
         temp_dq[:, :] = 0  # initialize for next integration
 
     # Average the slopes over all integrations.
-    # number_ints = data_sect.shape[0]
     if number_ints > 1:
         sum_weight = np.where(sum_weight <= 0., 1., sum_weight)
         recip_sum_weight = 1. / sum_weight
@@ -1270,7 +1267,6 @@ def compute_slope(
         cr_var_sect is a 3-D ndarray, the variance of each cosmic ray
         amplitude.
     """
-
     cr_flagged = np.empty(data_sect.shape, dtype=np.uint8)
     cr_flagged[:] = np.where(np.bitwise_and(gdq_sect, jump_flag), 1, 0)